+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Persistent expiration times. Expiration rejection jobs are rescheduled on startup.
 - Custom configuration variables are optional and override defaults.
 - Send messages to topics is optional.
+- Responses to bot can be sent to a different (than join requests) topic.
 
 ### Changed
 - Default configuration (mandatory) path is resolved at runtime (to the same directory as bot.py).